feat: start displaying a header with the version
The main reason for this is to ensure the user knows which version of
skan they are using. There actually isn't a great way to achieve this
with scala-cli so we hacked it in... Basically we need to ensure that
when a user pretty much runs anything that a script is ran ahead of that
to ensure that a `BuildInfo.scala` file is created with the information
that we need it in, like the version. For now since sbt-dynver isn't
cross published (the library part) I literally just copied that file
over. Once https://github.com/sbt/sbt-dynver/pull/245 is merged and
released I can just use that and delete everything I have locally making
that script quite small.

Closes #33