feat: add in an archive directory for deleted contexts
By default we used to just blow away a context when it was deleted.
Instead now we archive it by default. In the future we can probably
expand on this and choose to maybe bring back a context or add a setting
to really delete it instead of archiving it.